+++ src/001-050/S35.h
@@ -12,7 +12,7 @@ class S35 : public ISolution
 {
 public:
 	std::string GetAnswer() override;
-	std::vector<long long> GetRotations(long long i);
-	bool IsCircularPrime(long long prime, const std::vector<long long> &primes);
+	std::vector<int64_t> GetRotations(int64_t i);
+	bool IsCircularPrime(int64_t prime, const std::vector<int64_t> &primes);
 };
 } // namespace ProjectEuler
