+++ src/051-100/S72.cc
@@ -7,7 +7,7 @@ namespace ProjectEuler
 {
 std::string S72::GetAnswer()
 {
-	long long count = 0;
+	int64_t count = 0;
 
 	for (int d = 2; d <= 1000000; d++)
 	{
