+++ src/101-150/S115.cc
@@ -12,7 +12,7 @@ std::string S115::GetAnswer()
 	int n = 50;
 	while (true)
 	{
-		long long res = F(50, n);
+		int64_t res = F(50, n);
 		if (res > 1000000)
 		{
 			break;
@@ -24,14 +24,14 @@ std::string S115::GetAnswer()
 	return std::to_string(n);
 }
 
-long long S115::F(int m, int n)
+int64_t S115::F(int m, int n)
 {
 	int Layer = n;
 	int RedBlockLength = Layer;
 
 	// Color: grey[0] or red[1:RedBlockLength]; Layer
 	// dimensions: [RedBlockLength + 1][Layer]
-	// In C# it was new long long[RedBlockLength + 1, Layer]
+	// In C# it was new int64_t[RedBlockLength + 1, Layer]
 	// Access was reds[red, layer]
 	// So first dim is red length, second is layer index.
 
@@ -39,10 +39,10 @@ long long S115::F(int m, int n)
 	// In loop: layer < Layer. Access layer-1.
 	// reds[0, layer] = count.
 
-	// Let's use vector<vector<long long>> reds(RedBlockLength + 1, vector<long long>(Layer + 1, 0));
+	// Let's use vector<vector<int64_t>> reds(RedBlockLength + 1, vector<int64_t>(Layer + 1, 0));
 	// To be safe with indices.
 
-	vector<vector<long long>> reds(RedBlockLength + 1, vector<long long>(Layer + 1, 0));
+	vector<vector<int64_t>> reds(RedBlockLength + 1, vector<int64_t>(Layer + 1, 0));
 
 	// Original: reds[0, 0] = 1; reds[1, 0] = 1;
 	reds[0][0] = 1;
@@ -51,7 +51,7 @@ long long S115::F(int m, int n)
 
 	for (int layer = 1; layer < Layer; layer++)
 	{
-		long long count = 0;
+		int64_t count = 0;
 		for (int red = m; red <= RedBlockLength; red++)
 		{
 			count += reds[red][layer - 1];
@@ -66,7 +66,7 @@ long long S115::F(int m, int n)
 		}
 	}
 
-	long long result = 0;
+	int64_t result = 0;
 	for (int red = m; red <= RedBlockLength; red++)
 	{
 		result += reds[red][Layer - 1];
